# api_muebles/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Muebles",
        "description": "API REST para la gestión de muebles y categorías de Habita",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5502/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/categorias": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Listar todas las categorías",
                "description": "Endpoint público. Devuelve todas las categorías disponibles.",
                "operationId": "9bd4c5d463ba7b15843eebef947ed231",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Sofás"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Sofás y sillones"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crear una nueva categoría",
                "description": "Requiere token con ability `muebles.crear` (rol Administrador o Gestor).",
                "operationId": "33561abcf774ce0a7d668f231512afd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mesas"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Mesas de comedor y auxiliares"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Ver una categoría por ID",
                "description": "Endpoint público. Devuelve la categoría con sus muebles asociados.",
                "operationId": "0ddbcef186a80938088c691755aee7fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la categoría con sus muebles"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar una categoría",
                "description": "Requiere token con ability `muebles.editar` (rol Administrador o Gestor).",
                "operationId": "3cb64fbc825d55fb192dc610c8ce7a48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mesas de comedor"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar una categoría",
                "description": "Requiere token con ability `muebles.eliminar` (rol Administrador o Gestor).",
                "operationId": "5c8248a20636e938f3e6b20513b1d664",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Bearer obtenido desde la API de Usuarios al hacer login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}